[#143] Created Breath First Search
The search algorithm in combination with the previously created methods are all used in conjunction from this one method call. By having a graph with 2 valid points within it, this method will return an array of points corresponding to the shortest path. This also includes test code in master to test it