[#180] Changed the method names to more relevant names for other devs to understand the specific functionality.
There was an initial problem with reading external text files. The issue was that ShuttleInfo is not an activity/fragment class, hence it had no "context" and it could not actively get a resource, such as the external text files. This commit addresses that by passing down the context from the Maps Activity class, which allows us to read a resource file.