[#180] Using setter instead of assigning. Made the Context variable into a private static one, in order to prevent memory leaks and strengthen security. The latter will be accessed with getters and setters from the ShuttleInfo class.
There was an initial problem with reading external text files. The issue was that ShuttleInfo is not an activity/fragment class, hence it had no "context" and it could not actively get a resource, such as the external text files. This commit addresses that by passing down the context from the Maps Activity class, which allows us to read a resource file.